requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
